# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Pex Universe API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/v1/categories": {
            "get": {
                "description": "Get List of `Categories`",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "limit of results",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/routes.CategoriesResp"
                            }
                        }
                    }
                }
            }
        },
        "/v1/categories/{id}": {
            "get": {
                "description": "Get `Category` Info By ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Category ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/product.Category"
                        }
                    }
                }
            }
        },
        "/v1/categories/{id}/children": {
            "get": {
                "description": "Get List of `Categories` as Children of a `Category` with ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Category ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/product.Category"
                        }
                    }
                }
            }
        },
...
        "/v1/profile": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profile"
                ],
                "summary": "Get Profile Info",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/user.User"
                        }
                    }
                }
            },
            "put": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profile"
                ],
                "summary": "Update Profile",
                "parameters": [
                    {
                        "description": "Profile Update Dto",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routes.ProfileUpdateDto"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/user.User"
                        }
                    }
                }
            }
        },
        "/v1/profile/addresses": {
            "get": {
                "description": "Get List of `Address`es for the current `User`",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profile"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "limit of results",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/routes.AddressesResponse"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new `Address` for the current `User`",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profile"
                ],
                "parameters": [
                    {
                        "description": "Request Body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/address.AddressCreateDto"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/address.Address"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/profile/addresses/{id}": {
            "get": {
                "description": "Get `Address` Info By ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profile"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Address ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/address.Address"
                        }
                    }
                }
            },
            "put": {
                "description": "Update an `Address` by it's `ID`.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profile"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Address ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Request Body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/address.AddressUpdateDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/address.Address"
                        }
                    }
                }
            },
            "delete": {
                "description": "Get `Address` Info By ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profile"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Address ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/address.Address"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "address.Address": {
            "type": "object",
            "properties": {
                "city": {
                    "type": "string",
                    "example": "Tokyo"
                },
                "company": {
                    "type": "string"
                },
                "country": {
                    "$ref": "#/definitions/address.Country"
                },
                "createdAt": {
                    "type": "string"
                },
                "email": {
                    "type": "string",
                    "example": "john@example.com"
                },
                "ext": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string",
                    "example": "John"
                },
                "id": {
                    "type": "integer"
                },
                "lastName": {
                    "type": "string",
                    "example": "Doe"
                },
                "phone": {
                    "type": "string",
                    "example": "+11349503120"
                },
                "state": {
                    "$ref": "#/definitions/address.State"
                },
                "streetAddress1": {
                    "type": "string"
                },
                "streetAddress2": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "verified": {
                    "type": "boolean"
                },
                "zip": {
                    "type": "string"
                }
            }
        },
        "address.AddressCreateDto": {
            "type": "object",
            "properties": {
                "city": {
                    "type": "string",
                    "example": "Los Angeles"
                },
                "company": {
                    "type": "string"
                },
                "countryID": {
                    "type": "integer",
                    "minimum": 1
                },
                "email": {
                    "type": "string",
                    "example": "john@example.com"
                },
                "ext": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string",
                    "example": "John"
                },
                "lastName": {
                    "type": "string",
                    "example": "Doe"
                },
                "phone": {
                    "type": "string",
                    "example": "+12380941034"
                },
                "stateID": {
                    "type": "integer",
                    "minimum": 1
                },
                "streetAddress1": {
                    "type": "string"
                },
                "streetAddress2": {
                    "type": "string"
                },
                "zip": {
                    "type": "string"
                }
            }
        },
        "address.AddressUpdateDto": {
            "type": "object",
            "properties": {
                "city": {
                    "type": "string",
                    "example": "New York"
                },
                "company": {
                    "type": "string"
                },
                "countryId": {
                    "type": "integer",
                    "minimum": 1
                },
                "email": {
                    "type": "string",
                    "example": "john@example.com"
                },
                "ext": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string",
                    "example": "Jane"
                },
                "lastName": {
                    "type": "string",
                    "example": "Doe"
                },
                "phone": {
                    "type": "string",
                    "example": "+12380941034"
                },
                "stateId": {
                    "type": "integer",
                    "minimum": 1
                },
                "streetAddress1": {
                    "type": "string"
                },
                "streetAddress2": {
                    "type": "string"
                },
                "zip": {
                    "type": "string"
                }
            }
        },
        "address.Country": {
            "type": "object",
            "properties": {
                "displayName": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "position": {
                    "type": "integer"
                },
                "ppCode": {
                    "type": "string"
                }
            }
        },
        "address.State": {
            "type": "object",
            "properties": {
                "fullName": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "info": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "tax": {
                    "type": "number"
                }
            }
        },
        "model.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "'field' has failed the constraint: 'constraint' (vale: 'actual value')"
                },
                "status": {
                    "type": "integer",
                    "example": 400
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "product.Category": {
            "type": "object",
            "properties": {
                "brandAttributeCollapsed": {
                    "type": "boolean"
                },
                "brandAttributeEnabled": {
                    "type": "boolean"
                },
                "buttonContent": {
                    "type": "string"
                },
                "buttonEnable": {
                    "type": "boolean"
                },
                "buttonTitle": {
                    "type": "string"
                },
                "children": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/product.Category"
                    }
                },
                "createdAt": {
                    "type": "string"
                },
                "customCategory": {
                    "type": "boolean"
                },
                "defaultSorting": {
                    "type": "boolean"
                },
                "displayTitle": {
                    "type": "string"
                },
                "firstDescription": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "image": {
                    "type": "string"
                },
                "metaDescription": {
                    "type": "string"
                },
                "metaTitle": {
                    "type": "string"
                },
                "position": {
                    "type": "integer"
                },
                "priceAttributeCollapsed": {
                    "type": "boolean"
                },
                "priceAttributeEnabled": {
                    "type": "boolean"
                },
                "published": {
                    "type": "boolean"
                },
                "secondDescription": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "routes.AddressesResponse": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "currentPage": {
                    "type": "integer"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/address.Address"
                    }
                },
                "totalPages": {
                    "type": "integer"
                }
            }
        },
        "routes.CategoriesResp": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "currentPage": {
                    "type": "integer"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/product.Category"
                    }
                },
                "totalPages": {
                    "type": "integer"
                }
            }
        },
...
        "routes.ProfileUpdateDto": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string",
                    "example": "John Doe"
                }
            }
        },
        "user.User": {
            "type": "object",
            "properties": {
                "addresses": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/address.Address"
                    }
                },
                "createdAt": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
    }
}